requirements.txt
# ============================================================
# NOVA 2.0 — Python requirements
# ============================================================

# Reinforcement Learning
stable-baselines3==2.3.2
gymnasium==0.29.1

# Deep Learning / numerics
torch==2.3.0
numpy==1.26.4
pandas==2.2.2

# Config and utilities
PyYAML==6.0.2
matplotlib==3.9.2
tensorboard==2.17.1
tqdm==4.66.5

# SUMO Python bindings (includes traci + sumolib)
eclipse-sumo==1.24.0
